customer all apis tested and typeorm common usage

+++ apps/admin/src/user/user.service.ts
@@ -2,7 +2,7 @@ import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from 'common/entities/user.entity';
-import { RegisterDto } from '../auth/email.dto';
+import { RegisterDto } from '../auth-email/email.dto';
 import { UserUpdateDto } from './user.dto';
 import { MediaService } from '../../../../common/modules/media/media.service';
 import { plainToInstance } from 'class-transformer';
@@ -36,7 +36,7 @@ export class UserService {
     }
 
     await this.userRepository.update(user.id, updateData);
-    if (avatar && updateData.avatar_id !== previousImage) {
+    if (avatar && previousImage && updateData.avatar_id !== previousImage) {
       await this.mediaService.deleteMedia(previousImage);
     }
